Small scale refactoring
