+++ mypy.ini
@@ -27,9 +27,9 @@ strict_equality = True
 warn_unreachable = True
 warn_unused_ignores = True
 
-[mypy-lsst.daf.butler.*]
-ignore_missing_imports = False
-ignore_errors = False
+; [mypy-lsst.daf.butler.*]
+; ignore_missing_imports = False
+; ignore_errors = False
 
-[mypy-lsst.daf.butler.cli.*]
-ignore_missing_imports = True
+; [mypy-lsst.daf.butler.cli.*]
+; ignore_missing_imports = True
