+++ python/lsst/daf/butler_migrate/script/migrate_current.py
@@ -28,7 +28,7 @@ import logging
 
 from alembic import command
 
-from .. import config, migrate
+from .. import config, database
 
 
 _LOG = logging.getLogger(__name__)
@@ -50,11 +50,11 @@ def migrate_current(repo: str, mig_path: str, verbose: bool, butler: bool) -> No
         If True then print versions numbers from butler, otherwise display
         information about alembic revisions.
     """
-    db_url, schema = migrate.butler_db_params(repo)
+    db = database.Database.from_repo(repo)
 
     if butler:
         # Print current versions defined in butler.
-        manager_versions = migrate.manager_versions(db_url, schema)
+        manager_versions = db.manager_versions()
         if manager_versions:
             for manager, (klass, version, rev_id) in sorted(manager_versions.items()):
                 print(f"{manager}: {klass} {version} -> {rev_id}")
@@ -62,8 +62,5 @@ def migrate_current(repo: str, mig_path: str, verbose: bool, butler: bool) -> No
             print("No manager versions defined in butler_attributes table.")
     else:
         # Revisions from alembic
-        cfg = config.MigAlembicConfig.from_mig_path(mig_path)
-        cfg.set_main_option("sqlalchemy.url", db_url)
-        if schema:
-            cfg.set_section_option("daf_butler_migrate", "schema", schema)
+        cfg = config.MigAlembicConfig.from_mig_path(mig_path, db=db)
         command.current(cfg, verbose=verbose)
